Experiment to add a few more IPERF locations
Currently there is quite a bias towards Clouvider, and unfortunately, Clouvider seems to primarily rely on the transit of two networks: Arelion and GTT. This makes YABS IPERF3 output quite biased towards connectivity of these two providers. Would be nice if there could be a more diverse network selection, so I will be trying a few more options in this fork.